Move node stack parser to utils so it can be used from Electron
